final commit

+++ app/src/main/java/data/JSONWeatherParser.java
@@ -47,7 +47,7 @@ public class JSONWeatherParser {
             weather.currentCondition.setPressure(Utils.getInt("pressure",mainObj));
             weather.currentCondition.setMinTemp(Utils.getFloat("temp_min",mainObj));
             weather.currentCondition.setMaxTemp(Utils.getFloat("temp_max",mainObj));
-            weather.currentCondition.setTemperature(Utils.getDouble("temp",mainObj));
+            weather.currentCondition.setTemperature(Utils.getInt("temp",mainObj));
 
             JSONObject clodObj= Utils.getObject("clouds",jsonObject);
             weather.clouds.setPrecipitation(Utils.getInt("all",clodObj));
